Issue-3: add JSON example
