+++ profiles/anonintergroup/modules/custom/anonlocations/src/Entity/AnonLocation.php
@@ -183,7 +183,7 @@ class AnonLocation extends ContentEntityBase implements AnonLocationInterface {
       ->setLabel(t('Name'))
       ->setDescription(t('The name of the Anonymous 12 Step Location entity.'))
       ->setSettings(array(
-        'max_length' => 50,
+        'max_length' => 255,
         'text_processing' => 0,
       ))
       ->setDefaultValue('')
