+++ profiles/anonintergroup/modules/custom/anonmeetings/src/Entity/AnonMeeting.php
@@ -183,7 +183,7 @@ class AnonMeeting extends ContentEntityBase implements AnonMeetingInterface {
       ->setLabel(t('Name'))
       ->setDescription(t('The name of the Anonymous 12 Step Meeting entity.'))
       ->setSettings(array(
-        'max_length' => 50,
+        'max_length' => 255,
         'text_processing' => 0,
       ))
       ->setDefaultValue('')
