--- profiles/anonintergroup/modules/custom/anonmigrate/migration_templates/baltimoreaa-example-meetings.yml
@@ -1,35 +1,0 @@
-id: baltimoreaa_example_meetings
-label: Import Baltimore AA Example Meetings
-migration_tags: [ 'baltimoreaa-example' ]
-source:
-  plugin: anonintergroup
-  keys:
-    id: [ mAdd1, mAdd2, mCity, mZip ]
-    day: mDayNo
-  path: profiles/anonintergroup/modules/custom/anonmigrate/example-data/baltimoreaa.csv
-process:
-  name: mName
-  field_format:
-    plugin: meeting_format
-  field_location:
-    plugin: migration
-    migration: anonlocations
-    source: [ mAdd1, mAdd2, mCity, mZip ]
-  field_time/time:
-    plugin: meeting_time
-    source: mTime
-  field_time/length:
-    plugin: default_value
-    default_value: 60
-  field_time/sun: sun
-  field_time/mon: mon
-  field_time/tue: tue
-  field_time/wed: wed
-  field_time/thu: thu
-  field_time/fri: fri
-  field_time/sat: sat
-destination:
-  plugin: 'entity:anonmeeting'
-migration_dependencies:
- required:
-   - baltimoreaa_example_locations
